Ajout du module clear avec quelques modiffs

+++ modules/clear_module.py
@@ -12,7 +12,7 @@ logger = logging.getLogger('bot.clear_module')
 # Variables d'environnement pour la configuration
 moderator_role_ids_str = os.getenv('MODERATOR_ROLE_IDS', '')
 
-# Parsing les IDs de rôles modérateurs
+# Parsing des IDs de rôles modérateurs
 if moderator_role_ids_str:
     try:
         moderator_role_ids = list(map(int, moderator_role_ids_str.split(',')))
@@ -42,6 +42,10 @@ class ClearMessages(commands.Cog):
         
         await interaction.response.defer()
 
+        # Identifier l’ID du message de suivi
+        followup_message = await interaction.followup.send("Suppression des messages en cours...", wait=True)
+        followup_message_id = followup_message.id
+
         # Définir la limite de suppression des messages (14 jours)
         time_limit = datetime.now(timezone.utc) - timedelta(weeks=2)
         deleted_messages = 0
@@ -49,7 +53,8 @@ class ClearMessages(commands.Cog):
         if member:
             # Supprimer les messages d'un utilisateur spécifique
             def check(msg):
-                return msg.author.id == member.id and msg.created_at > time_limit
+                # Ne pas supprimer le message de suivi
+                return msg.author.id == member.id and msg.created_at > time_limit and msg.id != followup_message_id
 
             async for message in interaction.channel.history(limit=200).filter(check):
                 if deleted_messages < number:
@@ -60,13 +65,13 @@ class ClearMessages(commands.Cog):
         else:
             # Supprimer les derniers messages
             async for message in interaction.channel.history(limit=number):
-                if message.created_at > time_limit:
+                # Ne pas supprimer le message de suivi ou le message d'interaction ou les messages trop vieux
+                if message.id != interaction.id and message.created_at > time_limit and message.id != followup_message_id: 
                     await message.delete()
                     deleted_messages += 1
-                else:
-                    break
 
-        await interaction.followup.send(f"{deleted_messages} messages supprimés.")
+        # Editer le message de suivi avec le nombre de messages supprimés
+        await followup_message.edit(content=f"{deleted_messages} messages supprimés.")
 
 def setup(bot):
     bot.add_cog(ClearMessages(bot))
